+++ src/Accelerators/NNPA/CMakeLists.txt
@@ -25,7 +25,7 @@ add_subdirectory(Support)
 add_subdirectory(Transform)
 add_subdirectory(Compiler)
 
-add_onnx_mlir_library(NNPAAccel STATIC
+add_onnx_mlir_library(NNPAAccel
   NNPAAccelerator.cpp
 
   EXCLUDE_FROM_OM_LIBS
@@ -40,7 +40,7 @@ add_onnx_mlir_library(NNPAAccel STATIC
   LINK_LIBS PUBLIC
   onnx
   NNPACompilerUtils
-  # Accelerator
+  Accelerator
   OMZHighOps
   OMZLowOps
   )
