+++ src/Support/OMOptions.hpp
@@ -26,4 +26,3 @@ extern llvm::cl::opt<bool> enableMemoryBundling;
 extern llvm::cl::opt<int> onnxOpTransformThreshold;
 extern llvm::cl::opt<bool> onnxOpTransformReport;
 extern llvm::cl::opt<std::string> acceleratorTarget;
-extern llvm::cl::opt<onnx_mlir::OptLevel> OptimizationLevel;
